Avoid passing locale-dependent string to dateutil
When a date or datetime object is passed to freeze_time, it is
converted to a string which is then later parsed with dateutil.

The conversion was done using the '%x' format character, which return
the appropriate date representation for the current locale. This means
that you get '01/02/13' for January 2nd with the C and en_US locales,
and '02/01/13' for the same date using most European locales. The
string is then parsed incorrectly by dateutils.

Using the isoformat method on date/datetime objects solves this by
returning a string in 'yyyy-mm-dd' format. This string is parsed
correctly by dateutil, regardless of the current locale.