Fix `FakeTime.__call__()` to return the correct value when daylight savings is in effect.
The previous implementation was dependent on `mktime` which assumes the provided timetuple is in _local_ time.
The `time_to_freeze` is UTC, so `__call__` was attempting to compensate for this by using `time.timezone` to offset `time_to_freeze`.
The problem is that when daylight savings is in effect, `time.timezone` is wrong by an hour and consequently `__call__` returns a timestamp which is an hour early.

This issue can be addressed, and the code made much cleaner, by using simply `calendar.timegm` instead of `mktime`. `calendar.timegm` assumes the given time is in UTC so no timezone adjustment is needed and timestamp will always be correct, daylight savings or otherwise.