Hack for github/markup rendering SVG image badges
Github does not display well badges with SVG images as RST references. I've opened a pull to solve this bug, but the project does not seems maintained.

You can note that hovering PyPI version badge appends an underline after it like a link. Adding `#foo` after the link solves this.