Add support for perf_counter
`time.perf_counter` and `time.monotonic` are similar clocks (on many
platforms they are actually the *same* clock), with the main difference
between them being that if a system has a higher resolution monotonic
clock than the one that `time.monotonic()` calls, it will be used for
`time.perf_counter()`.

This commit adds support for `time.perf_counter()` and
`time.perf_counter_ns()` and refactors `monotonic()` and
`monotonic_ns()` so that these similar functions can share an
implementation.

It also refactors some of the tests around `monotonic()` and
`monotonic_ns()` so that the same tests can be used for
`perf_counter(_ns)` as well.