Restore class-decorator overload annotation for freeze_time
This was removed in f9235ae without explanation, possibly because mypy objects to the overlap under Python 3.13. But Pytype needs that to infer that the type is unchanged when a class is decorated with `@freezegun.freeze_time` (unlike mypy, which assumes class decorators, unlike function decorators, always leave the type unchanged).